requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.12.0
asteroid-filterbanks==0.4.0
asttokens==3.0.1
attrs==25.4.0
audioread==3.1.0
av==16.0.1
babel==2.17.0
bitarray==3.8.0
bitsandbytes==0.49.0
boto3==1.42.21
botocore==1.42.21
brotli==1.2.0
cached_path==1.8.1
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
coqpit-config==0.2.2
coqui-tts==0.27.3
coqui-tts-trainer==0.3.2
cycler==0.12.1
Cython==3.2.4
datasets==4.4.2
dateparser==1.1.8
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
docopt==0.6.2
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.9
encodec==0.1.1
executing==2.2.1
# Editable install with no version control (f5-tts==1.1.15)
-e /home/jonorl/Git/LLM-tutorial/F5-TTS
fairseq==0.12.2
faiss-cpu==1.13.2
fastapi==0.128.0
ffmpeg==1.4
ffmpy==1.0.0
filelock==3.19.1
fonttools==4.60.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gradio==6.2.0
gradio_client==2.0.2
greenlet==3.3.0
groovy==0.1.2
grpcio==1.76.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.0
inflect==7.5.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonlines==1.2.0
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
ko-speech-tools==0.1.0
lazy_loader==0.4
librosa==0.11.0
lightning==2.6.0
lightning-utilities==0.15.2
llvmlite==0.46.0
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mlxtend==0.23.4
monotonic-alignment-search==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.3
num2words==0.5.14
numba==0.63.1
numpy==2.3.5
omegaconf==2.3.0
openai-whisper==20250625
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optuna==4.6.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
portalocker==3.2.0
praat-parselmouth==0.4.7
primePy==1.3
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote-audio==4.0.3
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.3.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pypinyin==0.55.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytorch-lightning==2.6.0
pytorch-triton-rocm @ https://download.pytorch.org/whl/pytorch_triton_rocm-3.5.1-cp312-cp312-linux_x86_64.whl#sha256=7de1810936e253b18e2993fa50312357f7e28ed96b2cdc64d0bffe789cbf5ac5
pytz==2025.2
pyworld==0.3.5
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
rich==13.9.4
rjieba==0.1.13
rsa==4.9.1
s3transfer==0.16.0
sacrebleu==2.5.1
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
semantic-version==2.10.0
sentry-sdk==2.48.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
SQLAlchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.1+rocm6.4
torch-audiomentations==0.12.0
torch-tb-profiler==0.4.3
torch_pitch_shift==1.2.5
torchaudio==2.9.1+rocm6.4
torchcodec==0.9.1
torchcrepe==0.0.24
torchdiffeq==0.2.5
torchinfo==1.8.0
torchmetrics==1.8.2
torchvision==0.24.1+rocm6.4
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
transformers-stream-generator==0.0.5
triton==3.5.1
typeguard==4.4.4
typer==0.21.0
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.40.0
vocos==0.1.0
wandb==0.23.1
wcwidth==0.2.14
Werkzeug==3.1.4
x-transformers==2.12.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
